Change the behaviour of Text.GetChar and Text.SetChars when the TEXT parameter contains a character value outside of CHAR.  Formerly, they would "CHOP" the value, i.e., just zero out the high order bits of the value.
The new behaviour is to raise a range exception.

It lies beyond my imagination why anybody would want the old behaviour.
It would only make sense when the value is in CHAR, in which case, this
change will have no effect.  If you suddenly start getting these exceptions
from code that was formerly working, I modestly suggest that perhaps
it really wasn't.