TODO messages in UnsafeUniWr.m3.
One for putting TextForAllDo into Text, for avoiding
copying of characters and wider availability.

Another for adding <*FATAL ALL EXCEPT ... *> to compiler.
Without this, it is very difficult to avoid extraneous
warnings when a callback has a limited but nonempty exception
set, but the procedure doing the calling back must have ANY,
for generality.  This is a place where this happens.