Comments relevant to eventual new TipeDesc.Op.Widechar.
The values in TipeDesc.Op are used in runtime type descriptions
used by pickles.  A new operator was never added for WIDECHAR.
Instead, it is encoded as an enumeration type with a large
element count.  It would be nice to fix this someday.  There
are various occurrences of "(* Widechar Tipe. *)" marking known
places where coordinated changes would be needed.

 Changes to be committed:

	modified:   builtinTypes/WCharr.m3
	modified:   types/EnumType.m3