Rework (Fast) UniRd to utilize (Fast) UnGetCharMulti.
The previous scheme for handling necessary lookaheads in
UnsafeUniRd was written when there was no multi-character
unget provided by Rd.i3 for byte streams.  It used state
variables in the UniRd.T.  This meant sneaking around the
UniRd.T to its source would not work in any meaningful
way.  It also stood in the way of a reasonable implementation
of UniRd.UnGetCodePoint, which has never been implemented.

This change reworks the lookahead scheme to utilize Rd.UngetCharMulti
on the Source, fixing both problems.

Changes to be committed:
  modified:   UniRd.m3
  modified:   UniRdClass.i3
  modified:   UnsafeUniRd.m3