Fix assertion failure when waiting thread gets both Signaled and Alerted.
Using pthreads, in ThreadPThread.m3, if a thread were waiting in XWait, on a
M3 Condition variable, and it got both Signaled and Alerted, before XWait
reacquired the pthread mutex of the waiting thread's activation, an
overzealous assertion was failing.  This was showing up consistently
when formsedit tried to terminate.

 Changes to be committed:

	modified:   ThreadPThread.m3