More graceful and consistent handling of failures and their leftover files.
Consistently do:
  If a compilation phase fails, remove its output file,
    but only if -keep was not specified.
  If a compilation phase fails, don't run later phases.

Also, delete some redundant file remove actions in PushOneM3 that are
already done in the Run* procedure.  Some duplicate remove cases remain
when a failing phase is not the final phase, thus its output file
is also a temp file.  It is easier to leave this alone, as the Run*
procedures are also called from other places besides PushOneM3,
if a compilation starts with other than Modula-3 code.

And a small off-by-one bugfix in an assertion.

 Changes to be committed:
	modified:   Builder.m3