More/better command options for m3llvm: -g, -d.
The -d was intended to control trace output of m3llvm, for helping
debug itself.  Add a -g to control emission of debugger info into
the code begin translated, and set it as with similar options for
other backend modes.

Also, fix cm3 subcommand emission for m3llvm and llc.

 Changes to be committed:

	modified:   m3-sys/cminstall/src/config-no-install/Unix.common
	modified:   m3-sys/llvm3.6.1/src/M3CG_LLVM.i3
	modified:   m3-sys/llvm3.6.1/src/M3CG_LLVM.m3
	modified:   m3-sys/llvm3.6.1/src/Main.m3