Rename indistinguishable identifier "l" (letter ell).
In the font of my editor and terminal windows, "l", (letter ell)
and "1" (digit one) are completely indistinguishable.  Even when
magnified by xmag, one font is still totally indistinguishable and the
other just barely, with extremely close examination.

Since there are occurrences of "l+1", it's pretty hard to read.
So rename variable ell to "lab", systematically.