Put constructed name for gobal constant segment in its global variable.
The global constant segment comes from the front end with no name,
in the declare_segment operator.

M3CG_LLVM was already concocting a name for it ("M_Const"), but
not storing it in the variable for the segment, leading to a NIL
segfault later, when handling the end_init operator, in DebugGlobals,
while calling M3DIB.DIBcreateGlobalVariable.