Further try on FINALLY static links.  Also LOOPHOLE Fixes.
For most procedures, CG emits neither a static link formal nor
an actual for it in a call.  Backend must insert this stuff,
detecting whether the proc is nested, etc.

For an internally-generated FINALLY procedure, CG emits an explicit
formal for a static link, but does not emit an actual parameter
for in it a call.  CG does explicitly pass the SL to the runtime,
when pushing a finally frame, and this SL will be passed by the
runtime when it calls the FINALLY procedure.

This fixes execution runtime failures in p014, p035, p079, p140.

Also some fixes for LOOPHOLE, to conform to llvm's system of casts.