Infrastructure for passing static link rightmost (currently disabled.)
This may be useful in the future, but it did not fix the ABI
incompatability between m3llvm and m3cc backends.  And
putting it first will make life easier for a debugger,
probably a lot.  Leaving it in but disabled for now, i.e.,
static link is currently passed as leftmost parameter, as
it has been.