Fill in support for alloca.
The previous partial approach was wrong and inadequate.

alloca is a special case in a number of ways that
do not fit the general model/framework.

  - It's link name is __chkstk.
  - It takes its first and only parameter in eax.
  - It returns its return value in esp.

This lets the system mostly return to working.
You can bootstrap from 5.2.6, using so far an older
Visual C++ and get as far as the caltech lex/yacc rewrites.