Support for bootstrapping from the cm3-5.1.3 release, which must have been built with Visual C++ 4.0, or maybe 4.1 or 4.2. 4.0 in particular (4.1/4.2 unknown) output floating point division that requires Pentium fdiv workaround in the C runtime, that is missing in newer C runtimes.
Put this in sysutils instead of where I was previously
considering, a new LibcCompat package.

This is only needed for bootstrap, not for everything.
i.e. when using an old m3core.

The floating point division is in dtoa.c, which
maybe should be in Modula-3. "None" of the backends
generate this code -- well, the C backend with
old C compiler would.