Fix m3llvm NIL-derefs producing debug info for anonymous types.
m3llvm was suffering NIL derefs in Text, when producing debug
info for an anonymous type.  Change to give such a type a
name that is the base-62 form of its type UID, as used
ubiquitously in stabs debug info.  The fix is in many places,
hopefully all of them.