Add MDNodeFwdDec::replaceAllUsesWith(found in Metadata.h) to DIBuilder binding.
This is needed to resolve forward declarations of debug nodes,
which we need to create on account of various cycles.

This function eventually, in later versions of llvm, appears in
DIBuilder, but for now, we have to get it from Metadata.h.  I put
the binding in M3DIBuilder.*, because, aside from the above,
it uses types that are already there, and it's only one function.