In M3CG_LLVM, function DebugObject, extensive rework.
This function was a hastily cobbled-up mess that resulted from a
merge during the change to llvm-3.6.1, patched to just compile.
This change cleans it up greatly.  It necessarily uses the newly-
bound replaceAllUsesWith on MDNodes, because there is a very
local cycle in building object type debug info with a supertype.

More needs to be done with nonlocal cycles, predefined object types,
and the supertype of TEXT is questionable.  Also, DebugRecord is
a similar poorly merged mess.