Put "c" at the end of BUILD_DIR if using the C backend. This is controversial and convenient and annoying at the same time. It allows for easier coexistance of the gcc-based backend and the C backend. Most people don't care what they are using as long as it works reasonably, and if/when the C backend is the primary one, this will just be noise.
But while they are both around, it helps to keep them separate.
They are *not* link-compatible, at least when pointers to nested functions
are around.

The logic here needs to be fixed to avoid appending "c" twice.
And maybe the name should be different.
And we need to think about a C++ backend for efficient exception handling.