Change GrumpyUnix from .ms and .is to _m.s and _i.s, analogous to _i.o and _m.o.
Grumpy is gcc and automake -- they understand .s automatically to
mean run the assembler. gcc is easy to convince .ms is assembler
and we have a long history of doing that, so leave that alone.

This is perhaps only of hypothetical interest.
It depends how much "bootstrap" uses automake,
and how much it does so via cm3 or python.

Python can rename the files easily enough.

GrumpyUnix is not used since we have made do with regular Unix.
GrumpyUnix is actually reasonable -- there really isn't a downside to its names.