Add some support to Pickles for [un]pickling pseudopointers.
Since the entore pickle mechanism has no knowledge of static types
of references, but runs solely on their dynamic types, there can
be only one way of handling pseudopointers in the entire link
closure of a main program.

Allow registering a special to [un]pickle pseudopointers, detected
by their being an odd byte address.

In the absence of a registered special, [un]pickle them as integers,
with the usual size and endianness conversions.