Squench (all but one) compile warnings in libunicode package.
There is one warning remaining, in UnsafeUniWr.m3 that cannot be
removed without having a praqma like:

<* FATAL ANY EXCEPT Range , Failure , Alerted *>

This happens in various places when a general traverser procedure has
a nonempty RAISES set of its own, but also takes a callback procedure
as a parameter.  The latter has to have a formal type with RAISES ANY,
for the traverser to be useable in general.