Fix 1st problem in cm3ide reported by Wolfgang Keller <9103784@gmx.de>
This seems to have been a timing nondeterminacy between starting up of
the web server and display threads of cm3ide.  These were able to
happen concurrently, and when the display thread asks the web browser
to open a page first, there is no server to provide it, and the web
browser fails.  The failure was intermittent on my machine.

Surely this must have worked at one time.  I have never used or gotten
inside cm3ide before, so this is a bit speculative, but I am guessing
the change from user threads to pthreads may have altered which one
usually wins the race, making the latent bug show up.