Allow assigning and passing of procedures in predefined modules Word and Long.
This compiler fix allows assigning the procedures in predefined modules
Word and Long to a variable or passing to a formal parameter of appropriate
procedure type.  Previously, attempting to do this resulted in a code
generator failure message.

In m3core, there are already libraries Word and Long, whose only
apparent purpose could be to support this, yet the compiler could
not use them.  Add some explanation to the libraries, as this is
confusing.

Also, for those builtin procedures that still cannot be passed or
assigne (e.g. Info.InfoThisLine), give a meaningful compile-time
error message, instead of failing.