Fix wrong M3CG_Checks on code generated for Compare_Swap, I think.
The code produced by CG matches the declaration of Compare_Swap
in Atomic.ig, specifically, the 2nd parameter is passed by reference,
and can be altered.  Fix M3CG_Check to agree.

The comments in the code generators and M3CG_Ops.i3 suggest it is
passed by value, but this would not work correctly.  The back ends
need review for this operation, and are probably wrong in more ways
that one.

m3cc, in particular segfaults handling this code, if the stuff
in atomic is compiled.