Document what declare_param belongs to.
declare_local occurences in a procedure signature (i.e., following a
declare_procedure, not in the body) are not duplicated when the
signature for the same procedure is imported, via import_procedure and
its following occurences of declare_param.

This is important, because an imported signature, with its parameters,
can occur inside a procedure body, where declare_local can also occur.
Were it not so, a declare_local in this context would be ambiguous as
to whether it belonged to the most recent import_procedure or to the
current body.