Add a README file to test p004, explaining different warning order.
The m3cc and llvm versions of the compiler produce the same set of
warning messages but in a different order.  This is probably because
these back ends request different order of nested procedures from
the front end.  Little else would affect the front end.

This test has no explicitly coded nested procedures, but exception
handling generates internally produced nested procedures.

The added README file just explains this.

Probably the difference in compile output could be checked automatically,
but it's a priority thing.