Fix special-case passing of static link parameter.
Add the llvm 'nest' attribute to the static link parameter, both
for direct and indirect calls.  This causes llvm to pass this
parameter in the way the target ABI specifies for a static link.
Without this, everything was almost working except for execution
of a finally block when an exception propagated through it.  This
would probably have worked too, if the runtime were also compiled
with the llvm backend.  But RTExFrame passes a static link to user
code, and, if compiled by different back ends, these were using
different parameter conventions.

With this change, all test cases that pass using the gcc back end
also pass using the llvm backend, plus one more (p140).  Also, llvm
incorrectly shows a failure on p004.  This is only a reordering
of warning messages produced by the front end, which is asked to
unnest nested procedures in different order by these two back ends.

This is with mixing m3cc-compiled runtime and llvm-compiled test
cases, which bodes well for interoperability of these back ends.