Refactor CopyOpenArray for call site use.
Variable.CopyOpenArray generates prolog code to copy an open array
passed by VALUE.  Such copies are also needed at call sites for an
open array passed READONLY that is not a designator and/or not the
same type as the formal.

Change CopyOpenArray so it uses values on the stack rather than a
Variable.T, passed in.  This is so it can be used at call sites,
where no Variable.T is necessarily available.

This change does not add call site copying.  It is pure refactor
and generates identical CG intermediate code for the prolog copy.