Add command line support for disabling debugger output and extend optimization levels
Added support for -g0 to disable debugger information being built by the backend
this is primarily to enable developers to read the generated assembly and compare
output from cmc3g and llvm. The option is available for both gcc and clang.

Also added is support for -O0 -O1 -O2 -O3 optimisation levels which are passed to the
backend. O0 disables optimisation and the other levels are in line with gcc and llvm
llc and opt tools.