Document a peculiarity in CM3 IR that reflects call-site copy.
For parameters of struct types, (which could be any static sized
in-memory aggregate, but used only for records and arrays), only the
call site has IR operators that indicate that a copy should be made,
thus effectively passing by value.  There is no IR indication of
by-value passing at procedure entry, making it easy for a back end to
do call site copy and, at best, subtle to do it at procedure entry.

Document this is M3CG_Ops.i3.