Set mem_type to Struct as per the comment
Array, Record and (Set types greater than word size) generate pop_struct
operators for their calls by value. This results in a declare_param for
each of these types in the case of directly callable functions from imported procedures.
These params were being declared as Addr resulting in a type mismatch. Also the declaration in
the Interface was Addr but in the Module was Struct resulting in some ambiguity.