Add test p032, passing of array parameters.
Test passing of array parameters, fixed and open, in all three modes.

For READONLY mode, this is subtle.  According to Modula3, only
nondesignators passed by value in READONLY mode.  (Anything that
is assignable is passed by reference, and anything not assignable
is statically illegal.)  The only array nondesignators are constants
and function results, neither of which is aliasable, so they are
regarded as already being copies and passed by reference at the low
level.