Add computation of expression alignments.
To fix bugs passing packed values as parameters, there is a need
to know aligments of expressions.  Merely using the alignment of
the expression's type is sometimes not enough.  The type knows
nothing of the context in which a value occurs.  Particularly,
the alignment of a value of packed type depends on whether it is
an array element or record or object field, and, if so, what its
predecessor elements/fields are.

This commit purely adds computation of expression aligments, with
no uses thereof yet.