packedVars branch, initial commit.
A start toward fixing many compiler failures, error detection
failures, limitations on packed fields/elements, etc.

This started with the highly mistaken notion that only a few patches
here and there would be needed.  After months, this is clearly not
so.  It is time to get the many changes already made committed, but
there is more to do, so this is a new branch.

As of this commit, the front group recompiles itself twice and handles
some additional uses of packed types.  Currently, most fetches and stores
work.  A packed scalar field/element is handled if it does not cross a word
boundary, and a packed field/element that is a record or array is
handled if either the entire aggregate lies within one word, or it
is byte-aligned and every packed field/element within follows this rule
recursively.

There is more to do for subarrays, parameters, and constructors, at least.
There are also new regressions in the compiler test suite.  There are
several new test cases, but these are not commited as of now.

The concept of alignment in the front end has had to be expanded into
alignment of a type and alignment of a variable, which are sometimes
different.  Also in CG, an independent alignment system for postponed
computations has had to be expanded into three kinds.