Revert overly stingy record and array size limits.
Limits on record and array sizes were too low to be able to
compile TextLiteral.i3 on a 32-bit host machine.

Note: ArrayType.m3 and RecordType.m3 need to be reworked to use
TInt for compile-time size computations.  As is, a cross-compile
from 32-bit host to 64-bit target won't be able to handle very
large aggregates.  OTOH, this is unlikely to happen.