Fix runtime assignability errors in elements of static array constructors.
When an out-of-range ordinal value is an element of a static array constructor,
only a warning was being produced, as is usual, since this is a checked runtime
error.  For most other cases, there is also a runtime check in the generated
assign code.  For static constructors, no error of any kind was being
generated.  The constructor was just initialized in static storage with the
incorrect element value, out of type range and/or left-truncated, it it wouldn't
fit.

This fix generates an unconditional runtime error at any point
where an array constructor with any such element errors is used.

The previous commit was not adequate.

Still to do:  Extend this behavior to runtime assignability checks besides
ordinal ranges, and to record constructors.